Timer fix s->ms

+++ generic.h
@@ -80,7 +80,7 @@ typedef struct t_player {
     int viewDirection; // Direction the PC is facing, 0-359 degrees
     int health; // Three Health points, third hit kills
     uint8_t hasShot; // Has the player shot after sending the last clientState
-    double shotCooldown; // PCs can shoot 2/s
+    long shotCooldown; // PCs can shoot 2/s
     int isHost; // Is the player the host - 0: False, 1: True
     int playerNumber; // Player's number
     char* playerName; // Player's name
@@ -95,7 +95,7 @@ typedef struct t_game {
     int enemyCount; // Number of enemies in game
     int levelNumber; // Number of the game level, defines level parameters
     int enemyLimit; // How many enemies current level will spawn
-    double enemySpawnRate; // How quickly new enemies are spawned
+    long enemySpawnRate; // How quickly new enemies are spawned
     int enemyBaseSpeed; // Base speed for enemies
     int currentState; // 0: Waiting game, 1: inLobby, 2: inGame
     int maxPlayers; // Maximum amount of players, 1-4
